requirements.txt
fastapi
keras
Pillow
numpy
tensorflow
python-multipart
uuid0